requirements.txt
anthropic>=0.30.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
streamlit>=1.35.0
pytesseract>=0.3.10
pillow>=10.3.0
pdfplumber>=0.11.0
python-dotenv>=1.0.0
pydantic>=2.7.0
